Rename CQR and RetrievalPipeline
