+++ chatty_goose/cqr/__init__.py
@@ -1,1 +1,3 @@
+from .hqe import *
+from .ntr import *
 from .cqr import *
