+++ chatty_goose/pipeline/__init__.py
@@ -1,0 +1,1 @@
+from .retrieval_pipeline import *
