--- chatty_goose/retrievers/__init__.py
@@ -1,3 +1,0 @@
-from .hqe import *
-from .ntr import *
-from .retriever import *
